JuaLearn/jualearn_backend_root/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt
djoser
drf-nested-routers
Django
djangorestframework
pytest-django
pytest
